Fix two bugs in the price tag feature
"The price tag feature currently has two bugs.
1. The final price that a card ends up with is the one in the last set
alphabetically, instead of the one from the set with the lowest price.
2. Black Lotus Project sometimes gets in prices for cards from Masters
Edition, for some reason. This set only exists in Magic the Gathering:
Online, i.e., not physically, and should not be considered. This might
be considered a BLP bug and not a Cockatrice bug.

This patch fixes those two bugs. I also changed the price value to be
based on BLP's "price" value instead of the "average" value. Seems more
logical."

Signed-off-by: Daenyth <Daenyth+git@gmail.com>