bug fix: move isEnabled test out of acceptCardAttr
in the case of AND'ed filter terms, returning true
from acceptCardAttr will be equivalent to ignoring the
disabled term, but in the case of OR'ed terms,
returning true will incorrectly cause all cards
to be accepted.

the fix is to test for whether the term is enabled
before calling acceptCardAttr and continue if
disabled.