Updated Search Bar Style
Please note that I don't feel that the search bar needs to be skinned by
users, hence why it is not part of the large css-skinning project. I
also feel that given the PR #416 a new search bar with highlight the new
search functionality.

Updated the look of the search  bar:
+ larger, curved edges
+ changes color when has focus
+ search icon built into search
- removed "Search For" text in lieu of search icon