Fixed 5+ hour issue for new games
The problem was converting a negative int to an unsigned int. This was
resulting in a value of: (2^32) - n. Meaning that the room age was
extreamly old. After 48s the delay from the server will catch up and the
time diff will be 0 rather than negative, which is why after 48s the
room will change from 5+ hours to <1 min.