Theme manager - first part
This is the first implementation of a theme manager.
By now it replaces the previous options for hand,table,stack,player
backgrounds.
The “zonebg” folder has been removed in favor of a “themes” folder
where files have been reorganized in 5 different themes (one folder per
theme).
The first time cockatrice will run with this patch, is will copy themes
in place and setup the default theme.
This is the first part of the full implementation of a theme manager,
that have been split to avoid making a giant, very hard to review
commit.
Next parts will gradually include in the theme all the images and
styles used in the UI in cockatrice.