Fixed token PT issue for #931
The problem was that back when I added #336 I made it so that if YOU
didnt recive the PT of the token, then it would be fetched from your
tokens.xml, the issue is that we needed to send that value to the
server. The server sees it as a 0/0

This fix will send the PT from the tokens.xml if available.