Previous server connection history
This adds the ability to select previous servers from a list to connect
to.

You can remove items from the drop down by selecting them and pressing
delete.

If you connect to a new host it will be added to the previous hosts.

It will remember the last host connected to in the dropdown.