Add donator recognition
Added server side functionality to allow recognition of user accounts
that have donated. If the user is marked as a donator they are granted
server access regardless of maximum user settings.  In doing so I have
moved the max user check out of the server socket interface and moved it
into the protocol handler (were it should live honestly) and also
created a response to the server login request instead of triggering an
event for the server capacity being reached.