open decks and replays from file system
uses qlocalserer and qlocalpeer with filelocks to make clicked links in file system open in the active cockatrice instance. requires the mime type to be installed to work. linux only at the moment as windows requires registry changes to get it to be recognised as the default application for .cor/.cod files.

code cleanup