Introduce feature-based alternative events
This patch introduces a new feature into the Cockatrice protocol, which
I call alternative events. The problem goes like this: you want to
develop a cool new feature in the server, but you don't want to break
backwards-compatibility for clients that don't support the feature.
Sometimes the feature can be carefully designed so that clients that
don't support the new events and/or fields still behave reasonably.
Other times doing so is impossible or too complex without degrading
functionality, in which case the only current option is to introduce a
new required feature or accept that old clients will exhibit broken
behaviors.

Alternative events offer an intermediate solution to this issue. The
idea is to allow sending multiple versions -- upgrades -- of an event
stream (specifically, of a GameEventContainer), where each version is
tagged with a set of required features. Depending on its own features, a
client can select the appropriate event list to handle, and discard the
others.

In order to ensure backwards-compatibility of the alternative events
feature itself, it is implemented by making the existing
GameEventContainer's event_list the default event list for clients to
use, and indicate alternative events in a new repeated field. This way,
old clients not aware of the feature will use the default event list,
which is the expected behavior for a client that supports none of the
necessary features.

The changes in the client are fairly simple: when processing a
GameEventContainer, instead of using the event_list, we iterate over all
the upgrades and keep the latest one we have all the required features
for.

On the server, the changes are mostly centered around the
GameEventStorage. The GameEventStorage now delegates the construction of
GameEvent and GameEventContainer messages to a GameEventsBuilder, which
is responsible for building alternative event lists.