Use std::shared_ptr for Server_CardZone
On the server side, zones are owned by the player, and are deleted when
the player is deleted.  In turn, the zone owns its constituent cards,
and the zone pointer in the card is effectively a weak pointer.

This patch is preliminary work for the introduction of dynamic zones. It
makes these relationships between player, zones and cards exlicit by
using std::shared_ptr and std::weak_ptr instead of naked pointers for
Server_CardZone, hopefully limiting the risks of introducing memory bugs
such as leaks when implementing dynamic zones.