Introduce attach="transform" on <related> and <reverse-related>
This patch introduces a new value, "transform", for the attach=""
attribute on <related> and <reverse-related>. This is presented to the
user using a "Token: Transform to XXX" menu entry instead of the
existing "Token: Attach to XXX" menu entry.

When attach="transform" is specified, instead of attaching the card to
the newly created token, the original card disappears and is replaced by
the token. When the token changes zones, it is destroyed, and the
original card is moved to the same zone the token was moved into before
being destroyed.

This behavior is implemented by creating a dynamic zone attached to the
token, and moving the original card to that dynamic zone. If the server
does not support dynamic zones, the behavior is the same as
`attach="attach"` instead.