Support dynamic zones attached to a card
This patch gives Cockatrice a new feature: the ability to create new
zones during the game and attach them to a card.

This is used to implement a solution for #4081: to transform a card, we
create a token to represent the other face, and attach it to a new zone.
The new card is then moved to this new zone. When the token is
destroyed, the attached zone is automatically destroyed, and its content
move to wherever the token would have gone had it not been destroyed.

This patch is mostly concerned with server changes: the client is only
updated to deal with the new CreateZone and DestroyZone events, but no
functionality is provided to actually make use of the feature in the UI.

Note that current Cockatrice clients do not support MoveCard events
involving dynamic zones, and will silently discard them. This will be
dealt with in a later patch.

One relatively big part of this patch is refactors the helper function
Server_Player::moveCard to remove some of its responsibilities (namely,
parsing CardToMove protobufs into Server_Card objects, and deciding
whether a card move is legal or not --- both of which are now done by
the command that actually uses the helper). Some of the work is
extracted to other functions such as
Server_Game::removeArrowsRelatedToCard,
Server_Game::unattachCardsRelatedToCard, the [visible] parameter to
Server_CardZone::insertCard, and Server_Player::destroyCard.