fix: Always prefer local cards if available
== The Problem

Cockatrice has two ways of loading user-defined pictures from disk:

 - Through the CUSTOM directory, which ignores set information (any
   `Forest.png` file will be used for a `Forest`)
 - Through set-based directories in the pics folder (e.g.
   `M20/Forest.png` and `10E/Forest.png`) which get loaded depending on
   the user-defined set preference order

(There is also the legacy `downloadedPics` folder for the old picture
cache, which works like the set-based directories and is kept for
backwards compatibility only)

With recent provider-related changes, we only look for pictures on disk
for the most preferred set of a card (unless the "override all card art
with personal preference" setting is set). This makes sense for the
CUSTOM directory, but breaks the set-based directories, as they are
effectively not used anymore.

== The Solution

This patch changes the default logic for loading cards to always look
for cards of the requested set on disk, but to only look in the CUSTOM
folder for the most preferred set of a card. If the "override all card
art with personal preference" setting is set, always look in the CUSTOM
folder.

Note that this means that for sets that have multiple printings of a
card, and the user has a card for that set in their pics directory, the
same picture will be used for all printings in the set (which I think is
desired behavior).

=== Old Behavior

 - With "override all card art with personal preference" disabled
   (default): load from the CUSTOM directory and set-based directories
   for the user's most preferred set. Ignore the CUSTOM directory and
   set-based directories for other sets. Use provider set order.

 - With "override all card art with personal preference" enabled: load
   from the CUSTOM directory and set-based directories for the user's
   most preferred set. Ignore the CUSTOM directory and set-based
   directories for other sets. Use set order from settings.

=== New Behavior

 - With "override all card art with personal preference" disabled
   (default): load from the CUSTOM directory for the user's most
   preferred set, then always load from set-based directories. Use
   provider set order.

 - With "override all card art with personal preference" enabled: always
   use the CUSTOM directory and set-based directory. Use set order from
   settings.